[fix]: restore previous pwd
If we start a vim session and directly issue `:Zet` to edit a note, when
we close the buffer, the original pwd is lost since internally a `:lcd`
is used to switch to the notes dir. The solution is to issue the `:lcd
-` command whenever we close a notes buffer. Additionally, the
`BufWinEnter` autocmd is used instead of `BufNewFile,BufRead` since
`:lcd` only works for windows (not tabs). `BufWinLeave` is used to issue
the `:lcd -` command when we close a notes buffer.